fix: integration tests race and SmallFileThreshold overflow
- Use WaitForAllUploads instead of WaitForEagerUploads in integration
  tests since 500ms coalescing cooldown prevents eager uploads during
  rapid test writes
- Fix uint64-to-int64 overflow in SmallFileThreshold conversion by
  clamping before casting

+++ cmd/dfs/commands/start.go
@@ -133,7 +133,7 @@ func runStart(cmd *cobra.Command, args []string) error {
 		ParallelUploads:    cfg.Offloader.ParallelUploads,
 		ParallelDownloads:  cfg.Offloader.ParallelDownloads,
 		PrefetchBlocks:     cfg.Offloader.PrefetchBlocks,
-		SmallFileThreshold: min(int64(cfg.Offloader.SmallFileThreshold), math.MaxInt64),
+		SmallFileThreshold: int64(min(uint64(cfg.Offloader.SmallFileThreshold), uint64(math.MaxInt64))),
 	})
 	logger.Info("Cache configuration stored", "path", cfg.Cache.Path, "size", cfg.Cache.Size, "max_pending_size", cfg.Cache.MaxPendingSize)
